refactor: add hook to hide messages on profile page
