+++ src/components/Profile/Tabs/SecondTab/SecondTab.tsx
@@ -1,6 +1,6 @@
 import styles from './secondTab.module.scss';
 import { Formik } from 'formik';
-import { MouseEvent, ReactElement, useContext, useEffect, useState } from 'react';
+import { MouseEvent, ReactElement, useState } from 'react';
 import { Button } from '../../../shared/Button';
 import { initialEditAddresses } from '../../../../constant';
 import { CustomerAddress } from './CustomerAdress';
@@ -9,7 +9,7 @@ import classNames from 'classnames';
 import { IAddressForm } from '../../../RegisterForm/AddressForm/AddressForm';
 import { Radiobtn } from './Radiobtn';
 import { useSelector } from 'react-redux';
-import { selectEditUserInfo, selectUserData } from '../../../../store/user/selectors';
+import { selectUserData } from '../../../../store/user/selectors';
 import { useAppDispatch } from '../../../../store/store';
 import { Address } from '@commercetools/platform-sdk';
 import { IAddNewAddress, IRemoveAddress, ISetDefaultAddress } from '../../../../services/sdk/customer/types';
@@ -17,8 +17,7 @@ import { removeAddressThunk } from '../../../../store/user/thunks';
 import { setDefaultAddressIdThunk } from '../../../../store/user/thunks';
 import { addNewAddressThunk } from '../../../../store/user/thunks';
 import { ErrorMessage } from '../../../shared/ErrorMessage';
-import { deleteSuccessState, resetEditError } from '../../../../store/user/slice';
-import { ModeContext } from '../../../../context/mode/ModeContext.ts';
+import { useProfileMessages } from '../../../../hooks';
 
 export interface IAddressesProfile {
   shipping: IAddressForm;
@@ -27,38 +26,9 @@ export interface IAddressesProfile {
 
 function SecondTab(): ReactElement {
   const [isShipping, setIsShipping] = useState(true);
-  const { isEditMode, toggleEditMode } = useContext(ModeContext);
+  const [editStatus, editError, isSuccess, isEditMode, toggleEditMode] = useProfileMessages();
   const user = useSelector(selectUserData);
   const dispatch = useAppDispatch();
-  const { editStatus, editError, isSuccess } = useSelector(selectEditUserInfo);
-
-  useEffect(() => {
-    if (editStatus === 'success') {
-      if (isEditMode) {
-        toggleEditMode();
-      }
-
-      const timer = setTimeout(() => {
-        dispatch(deleteSuccessState());
-      }, 3000);
-
-      return () => {
-        clearTimeout(timer);
-      };
-    }
-  }, [dispatch, editStatus]);
-
-  useEffect(() => {
-    if (editError) {
-      const timer = setTimeout(() => {
-        dispatch(resetEditError());
-      }, 3000);
-
-      return () => {
-        clearTimeout(timer);
-      };
-    }
-  }, [dispatch, editError]);
 
   function handleSubmit(values: IAddressesProfile): void {
     const isShippingAddress = values.shipping.city !== '';
